+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [next_version] - 2020-02-03
+## [NEXT_VERSION]
 ### Changed
 - Added dnsConfig on deploy