+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [NEXT_VERSION]
+## [0.29.1-rc.2] - 2020-02-04
 ### Changed
 - Added dnsConfig on deploy